# server/src/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/roles": {
      "get": {
        "operationId": "RolesController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Role"
                          }
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get list of roles",
        "tags": [
          "Roles"
        ]
      }
    },
    "/v1/roles/{id}": {
      "get": {
        "operationId": "RolesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "$ref": "#/components/schemas/Role"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get role detail",
        "tags": [
          "Roles"
        ]
      }
    },
    "/v1/users": {
      "post": {
        "operationId": "UsersController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/employees": {
      "post": {
        "operationId": "EmployeesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "$ref": "#/components/schemas/Employee"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Create new employee with user data",
        "tags": [
          "Employees"
        ]
      },
      "get": {
        "operationId": "EmployeesController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Employee"
                          }
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get list of employees",
        "tags": [
          "Employees"
        ]
      }
    },
    "/v1/employees/{id}": {
      "get": {
        "operationId": "EmployeesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "$ref": "#/components/schemas/Employee"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get employee detail with attendances",
        "tags": [
          "Employees"
        ]
      },
      "put": {
        "operationId": "EmployeesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "$ref": "#/components/schemas/Employee"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Update employee",
        "tags": [
          "Employees"
        ]
      },
      "delete": {
        "operationId": "EmployeesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "$ref": "#/components/schemas/Employee"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Delete employee cascading",
        "tags": [
          "Employees"
        ]
      }
    },
    "/v1/attendances": {
      "post": {
        "operationId": "AttendancesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "$ref": "#/components/schemas/Attendance"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Log attendance",
        "tags": [
          "Attendances"
        ]
      },
      "get": {
        "operationId": "AttendancesController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Attendance"
                          }
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get list of attendances",
        "tags": [
          "Attendances"
        ]
      }
    },
    "/v1/attendances/{id}": {
      "get": {
        "operationId": "AttendancesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "$ref": "#/components/schemas/Attendance"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseWrapper"
                    },
                    {
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Error"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "logId": {
                          "type": "string",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get attendance detail",
        "tags": [
          "Attendances"
        ]
      }
    }
  },
  "info": {
    "title": "API Documentation",
    "description": "API Documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://172.20.5.250:3000",
      "description": "Local Server Dev"
    },
    {
      "url": "http://localhost:3000",
      "description": "Localhost"
    }
  ],
  "components": {
    "schemas": {
      "ResponseWrapper": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {}
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "Employee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {}
      },
      "Attendance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "CreateAttendanceDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}